+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Made it possible to construct a step graph out of `Step` objects, instead of constructing it out of `StepStub` objects.
+- Removed dataset fingerprinting code, since we can now use `Step` to make sure things are cached.
+- Made `Executor` non-registrable. This is a temporary state and will be changed back.
+- Made steps deterministic by default.
+- Brought back `MemoryStepCache`, so we can run steps without configuring anything.
 
 ## [v0.3.5](https://github.com/allenai/tango/releases/tag/v0.3.5) - 2021-11-05