+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added a `.log_batch()` method on `torch::TrainCallback` which is given the average loss across
+  distributed workers, but only called every `log_every` steps.
+
 ### Changed
 
 - Made it possible to construct a step graph out of `Step` objects, instead of constructing it out of `StepStub` objects.
@@ -14,6 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made `Executor` non-registrable. This is a temporary state and will be changed back.
 - Made steps deterministic by default.
 - Brought back `MemoryStepCache`, so we can run steps without configuring anything.
+
+### Removed
+
+- Removed `.pre_log_batch()` method on `torch::TrainCallback`.
+
+### Fixed
+
+- Fixed typo in parameter name `remove_stale_checkpoints` in `TorchTrainStep` (previously was `remove_state_checkpoints`).
 
 ## [v0.3.5](https://github.com/allenai/tango/releases/tag/v0.3.5) - 2021-11-05
 
@@ -26,13 +39,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `StopEarlyCallback`, a `TorchTrainCallback` for early stopping.
+- Added `StopEarlyCallback`, a `torch::TrainCallback` for early stopping.
 - Added parameter `remove_stale_checkpoints` to `TorchTrainStep`.
 
 ### Changed
 
-- Minor changes to `TorchTrainCallback` interface.
-- Weights & Biases `TorchTrainCallback` now logs best validation metric score.
+- Minor changes to `torch::TrainCallback` interface.
+- Weights & Biases `torch::TrainCallback` now logs best validation metric score.
 
 ## [v0.3.3](https://github.com/allenai/tango/releases/tag/v0.3.3) - 2021-11-04